refactor(payment): PAYPAL-2813 created BraintreeAcceleratedCheckoutUtils class
