+++ packages/braintree-integration/src/braintree-accelerated-checkout/braintree-accelerated-checkout-customer-strategy.spec.ts
@@ -3,41 +3,26 @@ import { getScriptLoader } from '@bigcommerce/script-loader';
 import {
     CustomerInitializeOptions,
     InvalidArgumentError,
-    MissingDataError,
     PaymentIntegrationService,
 } from '@bigcommerce/checkout-sdk/payment-integration-api';
-import {
-    getBillingAddress,
-    getCountries,
-    getCustomer,
-    PaymentIntegrationServiceMock,
-} from '@bigcommerce/checkout-sdk/payment-integrations-test-utils';
+import { PaymentIntegrationServiceMock } from '@bigcommerce/checkout-sdk/payment-integrations-test-utils';
 import { BrowserStorage } from '@bigcommerce/checkout-sdk/storage';
 
-import { BraintreeConnect } from '../braintree';
 import BraintreeIntegrationService from '../braintree-integration-service';
 import BraintreeScriptLoader from '../braintree-script-loader';
-import {
-    getBraintreeAcceleratedCheckoutPaymentMethod,
-    getBraintreeConnectProfileDataMock,
-    getConnectMock,
-} from '../mocks/braintree.mock';
 
 import BraintreeAcceleratedCheckoutCustomerStrategy from './braintree-accelerated-checkout-customer-strategy';
+import BraintreeAcceleratedCheckoutUtils from './braintree-accelerated-checkout-utils';
 
 describe('BraintreeAcceleratedCheckoutCustomerStrategy', () => {
-    let braintreeConnectMock: BraintreeConnect;
+    let braintreeAcceleratedCheckoutUtils: BraintreeAcceleratedCheckoutUtils;
     let braintreeIntegrationService: BraintreeIntegrationService;
     let braintreeScriptLoader: BraintreeScriptLoader;
     let browserStorage: BrowserStorage;
     let paymentIntegrationService: PaymentIntegrationService;
     let strategy: BraintreeAcceleratedCheckoutCustomerStrategy;
 
-    const countries = getCountries();
-    const customer = getCustomer();
-    const billingAddress = getBillingAddress();
     const methodId = 'braintreeacceleratedcheckout';
-    const paymentMethod = getBraintreeAcceleratedCheckoutPaymentMethod();
 
     const initializationOptions = { methodId };
     const executionOptions = {
@@ -46,8 +31,6 @@ describe('BraintreeAcceleratedCheckoutCustomerStrategy', () => {
     };
 
     beforeEach(() => {
-        braintreeConnectMock = getConnectMock();
-
         browserStorage = new BrowserStorage('braintree-accelerated-checkout-mock');
         braintreeScriptLoader = new BraintreeScriptLoader(getScriptLoader(), window);
         braintreeIntegrationService = new BraintreeIntegrationService(
@@ -55,40 +38,26 @@ describe('BraintreeAcceleratedCheckoutCustomerStrategy', () => {
             window,
         );
         paymentIntegrationService = new PaymentIntegrationServiceMock();
-
-        strategy = new BraintreeAcceleratedCheckoutCustomerStrategy(
+        braintreeAcceleratedCheckoutUtils = new BraintreeAcceleratedCheckoutUtils(
             paymentIntegrationService,
             braintreeIntegrationService,
             browserStorage,
         );
 
-        jest.spyOn(browserStorage, 'setItem').mockImplementation(jest.fn);
+        strategy = new BraintreeAcceleratedCheckoutCustomerStrategy(
+            paymentIntegrationService,
+            braintreeAcceleratedCheckoutUtils,
+        );
 
         jest.spyOn(paymentIntegrationService, 'loadPaymentMethod');
-        jest.spyOn(paymentIntegrationService, 'updatePaymentProviderCustomer').mockImplementation(
-            jest.fn,
-        );
-        jest.spyOn(paymentIntegrationService.getState(), 'getCustomer').mockReturnValue(customer);
-        jest.spyOn(paymentIntegrationService.getState(), 'getCountries').mockReturnValue(countries);
-        jest.spyOn(paymentIntegrationService.getState(), 'getBillingAddress').mockReturnValue(
-            billingAddress,
-        );
-        jest.spyOn(paymentIntegrationService.getState(), 'getPaymentMethodOrThrow').mockReturnValue(
-            paymentMethod,
-        );
-
-        jest.spyOn(braintreeIntegrationService, 'initialize');
-        jest.spyOn(braintreeIntegrationService, 'getBraintreeConnect').mockImplementation(
-            () => braintreeConnectMock,
-        );
-
-        jest.spyOn(braintreeConnectMock.identity, 'lookupCustomerByEmail').mockReturnValue({
-            customerId: 'customerId',
-        });
-        jest.spyOn(braintreeConnectMock.identity, 'triggerAuthenticationFlow').mockReturnValue({
-            authenticationState: 'succeeded',
-            profileData: getBraintreeConnectProfileDataMock(),
-        });
+        jest.spyOn(
+            braintreeAcceleratedCheckoutUtils,
+            'initializeBraintreeConnectOrThrow',
+        ).mockImplementation(jest.fn);
+        jest.spyOn(
+            braintreeAcceleratedCheckoutUtils,
+            'authenticatePayPalConnectUserOrThrow',
+        ).mockImplementation(jest.fn);
     });
 
     afterEach(() => {
@@ -104,60 +73,16 @@ describe('BraintreeAcceleratedCheckoutCustomerStrategy', () => {
             }
         });
 
-        it('throws an error if clientToken is not defined', async () => {
-            jest.spyOn(
-                paymentIntegrationService.getState(),
-                'getPaymentMethodOrThrow',
-            ).mockReturnValue({
-                ...paymentMethod,
-                clientToken: undefined,
-            });
-
-            try {
-                await strategy.initialize(initializationOptions);
-            } catch (error) {
-                expect(error).toBeInstanceOf(MissingDataError);
-            }
-        });
-
-        it('throws an error if initializationData is not defined', async () => {
-            jest.spyOn(
-                paymentIntegrationService.getState(),
-                'getPaymentMethodOrThrow',
-            ).mockReturnValue({
-                ...paymentMethod,
-                initializationData: undefined,
-            });
-
-            try {
-                await strategy.initialize(initializationOptions);
-            } catch (error) {
-                expect(error).toBeInstanceOf(MissingDataError);
-            }
-        });
-
-        it('initializes braintree integration service and loads braintree connect', async () => {
+        it('initializes braintree connect', async () => {
             await strategy.initialize(initializationOptions);
 
-            expect(braintreeIntegrationService.initialize).toHaveBeenCalledWith(
-                paymentMethod.clientToken,
-                paymentMethod.initializationData,
-            );
-            expect(braintreeIntegrationService.getBraintreeConnect).toHaveBeenCalled();
+            expect(
+                braintreeAcceleratedCheckoutUtils.initializeBraintreeConnectOrThrow,
+            ).toHaveBeenCalledWith(methodId);
         });
     });
 
     describe('#executePaymentMethodCheckout()', () => {
-        it('throws an error if methodId is not provided', async () => {
-            await strategy.initialize(initializationOptions);
-
-            try {
-                await strategy.executePaymentMethodCheckout();
-            } catch (error) {
-                expect(error).toBeInstanceOf(InvalidArgumentError);
-            }
-        });
-
         it('throws an error if continueWithCheckoutCallback is not provided or it is not a function', async () => {
             await strategy.initialize(initializationOptions);
 
@@ -176,141 +101,9 @@ describe('BraintreeAcceleratedCheckoutCustomerStrategy', () => {
         });
     });
 
-    describe('#authenticatePayPalConnectUser()', () => {
-        it('does not authenticate user if braintree connect is not loaded', async () => {
-            await strategy.executePaymentMethodCheckout(executionOptions);
-
-            expect(paymentIntegrationService.updatePaymentProviderCustomer).not.toHaveBeenCalled();
-        });
-
-        it('does not authenticate user if email is not defined', async () => {
-            const customerWithoutEmail = { ...customer, email: undefined };
-            const billingWithoutEmail = { ...billingAddress, email: undefined };
-
-            jest.spyOn(paymentIntegrationService.getState(), 'getCustomer').mockReturnValue(
-                customerWithoutEmail,
-            );
-            jest.spyOn(paymentIntegrationService.getState(), 'getBillingAddress').mockReturnValue(
-                billingWithoutEmail,
-            );
-
-            await strategy.initialize(initializationOptions);
-            await strategy.executePaymentMethodCheckout(executionOptions);
-
-            expect(paymentIntegrationService.updatePaymentProviderCustomer).not.toHaveBeenCalled();
-        });
-
-        it('checks if there is PP Connect account with customer email', async () => {
-            jest.spyOn(braintreeConnectMock.identity, 'lookupCustomerByEmail').mockReturnValue({});
-
-            await strategy.initialize(initializationOptions);
-            await strategy.executePaymentMethodCheckout(executionOptions);
-
-            expect(braintreeConnectMock.identity.lookupCustomerByEmail).toHaveBeenCalled();
-        });
-
-        it('does not authenticate user does not have PP Connect account', async () => {
-            jest.spyOn(braintreeConnectMock.identity, 'lookupCustomerByEmail').mockReturnValue({});
-
-            await strategy.initialize(initializationOptions);
-            await strategy.executePaymentMethodCheckout(executionOptions);
-
-            expect(paymentIntegrationService.updatePaymentProviderCustomer).not.toHaveBeenCalled();
-        });
-
-        it('triggers PP Connect authentication flow if customer is detected as PP Connect user', async () => {
-            await strategy.initialize(initializationOptions);
-            await strategy.executePaymentMethodCheckout(executionOptions);
-
-            expect(braintreeConnectMock.identity.triggerAuthenticationFlow).toHaveBeenCalled();
-            expect(paymentIntegrationService.updatePaymentProviderCustomer).toHaveBeenCalled();
-        });
-
-        it('successfully authenticate customer with PP Connect', async () => {
-            const updatePaymentProviderCustomerPayload = {
-                authenticationState: 'succeeded',
-                addresses: [
-                    {
-                        firstName: 'John',
-                        lastName: 'Doe',
-                        company: '',
-                        address1: 'Hello World Address',
-                        address2: '',
-                        city: 'Bellingham',
-                        stateOrProvince: 'WA',
-                        stateOrProvinceCode: 'WA',
-                        country: 'United States',
-                        countryCode: 'US',
-                        postalCode: '98225',
-                        phone: '',
-                        customFields: [],
-                    },
-                ],
-                instruments: [
-                    {
-                        bigpayToken: 'pp-vaulted-instrument-id',
-                        brand: 'VISA',
-                        defaultInstrument: false,
-                        expiryMonth: undefined,
-                        expiryYear: '02/2037',
-                        iin: '',
-                        last4: '1111',
-                        method: 'braintreeacceleratedcheckout',
-                        provider: 'braintreeacceleratedcheckout',
-                        trustedShippingAddress: false,
-                        type: 'card',
-                    },
-                ],
-            };
-
-            await strategy.initialize(initializationOptions);
-            await strategy.executePaymentMethodCheckout(executionOptions);
-
-            expect(braintreeConnectMock.identity.triggerAuthenticationFlow).toHaveBeenCalled();
-            expect(paymentIntegrationService.updatePaymentProviderCustomer).toHaveBeenCalledWith(
-                updatePaymentProviderCustomerPayload,
-            );
-            expect(browserStorage.setItem).toHaveBeenCalledWith('customer', {
-                email: customer.email,
-                ...updatePaymentProviderCustomerPayload,
-            });
-        });
-
-        it('does not authenticate customer if the authentication was canceled or failed', async () => {
-            jest.spyOn(braintreeConnectMock.identity, 'triggerAuthenticationFlow').mockReturnValue({
-                authenticationState: 'canceled',
-                profileData: {},
-            });
-
-            const updatePaymentProviderCustomerPayload = {
-                authenticationState: 'canceled',
-                addresses: [],
-                instruments: [],
-            };
-
-            await strategy.initialize(initializationOptions);
-            await strategy.executePaymentMethodCheckout(executionOptions);
-
-            expect(braintreeConnectMock.identity.triggerAuthenticationFlow).toHaveBeenCalled();
-            expect(paymentIntegrationService.updatePaymentProviderCustomer).toHaveBeenCalledWith(
-                updatePaymentProviderCustomerPayload,
-            );
-
-            expect(browserStorage.setItem).toHaveBeenCalledWith('customer', {
-                email: customer.email,
-                ...updatePaymentProviderCustomerPayload,
-            });
-        });
-    });
-
     describe('#deinitialize()', () => {
-        it('teardowns braintree sdk creator on strategy deinitialize', async () => {
-            braintreeIntegrationService.teardown = jest.fn();
-
-            await strategy.initialize(initializationOptions);
-            await strategy.deinitialize();
-
-            expect(braintreeIntegrationService.teardown).toHaveBeenCalled();
+        it('deinitializes strategy', async () => {
+            await expect(strategy.deinitialize()).resolves.not.toThrow();
         });
     });
 
@@ -339,8 +132,9 @@ describe('BraintreeAcceleratedCheckoutCustomerStrategy', () => {
             await strategy.initialize(initializationOptions);
             await strategy.signIn(credentials);
 
-            expect(braintreeConnectMock.identity.triggerAuthenticationFlow).toHaveBeenCalled();
-            expect(paymentIntegrationService.updatePaymentProviderCustomer).toHaveBeenCalled();
+            expect(
+                braintreeAcceleratedCheckoutUtils.authenticatePayPalConnectUserOrThrow,
+            ).toHaveBeenCalledWith(credentials.email);
         });
     });
